🚧 Modified entities and moved to domain folder
