+++ lib/features/reserve/presentation/widgets/info_widgets/climate_widget.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:tennis_court_app/features/reserve/infrastructure/entities/entities.dart';
+import 'package:tennis_court_app/config/config.dart';
+import 'package:tennis_court_app/features/reserve/domain/entities/entities.dart';
 
 class ClimateWidget extends StatelessWidget {
   const ClimateWidget({
@@ -12,19 +13,28 @@ class ClimateWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     double? rainProb = 0;
     if (weather != null) {
-      final rain = weather!.hourly!.firstWhere((element) {
-        rainProb = element.rain!.the1H;
+      final rain = weather!.data!.firstWhere((element) {
+        rainProb = element.rain;
         return rainProb! > 0;
       });
-      debugPrint("${rain.rain?.the1H}");
+      debugPrint("${rain.rain}");
     }
     return Row(
       children: [
-        SvgPicture.asset(
-          'assets/icons/rainy_icon.svg',
-          width: 16,
-          height: 16,
-        ),
+        if (weather?.data?.first.weather?.first.icon != null)
+          FadeInImage(
+              width: 16,
+              height: 16,
+              placeholder: const AssetImage("assets/images/no-image.png"),
+              image: NetworkImage(
+                "$iconsServerUrl/${weather!.data!.first.weather!.first.icon}@2x.png",
+              ))
+        else
+          SvgPicture.asset(
+            'assets/icons/rainy_icon.svg',
+            width: 16,
+            height: 16,
+          ),
         const SizedBox(width: 5),
         Text(
           rainProb != null ? "$rainProb" : 'N/A',
