+++ lib/features/reserve/reserve.dart
@@ -1,2 +1,3 @@
+export 'domain/domain.dart';
 export 'infrastructure/infrastructure.dart';
 export 'presentation/presentation.dart';
